begin refactor
